pyproject.toml
# [tool.poetry]
# package-mode = false

[project]
name = "paathguide"
version = "0.1.0"
description = ""
authors = [{ name = "ArmanGrewal007", email = "armansinghgrewal@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "alembic (>=1.16.5,<2.0.0)",
  "click (>=8.2.1,<9.0.0)",
  # "datasets (>=4.0.0,<5.0.0)",
  # "editdistance (>=0.8.1,<0.9.0)",
  # "einops (>=0.8.1,<0.9.0)",
  "fastapi (>=0.116.1,<0.117.0)",
  "isort (>=6.0.1,<7.0.0)",
  # "hydra-core (>=1.3.0,<2.0.0)",
  # "ipython (>=9.5.0,<10.0.0)",
  # "jiwer (>=4.0.0,<5.0.0)",
  # "jupyter (>=1.1.1,<2.0.0)",
  # "lhotse (>=1.0.0,<2.0.0)",
  # "librosa (>=0.11.0,<0.12.0)",
  # "lightning (>=2.0.0,<3.0.0)",
  "llvmlite (==0.44.0)",
  # "matplotlib (>=3.10.6,<4.0.0)",
  # "nemo-toolkit (>=2.4.0,<3.0.0)",
  # "notebook (>=7.4.5,<8.0.0)",
  # "numba (==0.61.2)",
  # "omegaconf (>=2.3.0,<3.0.0)",
  "openai-whisper (==20250625)",
  # "pandas (>=2.3.2,<3.0.0)",
  # "pyannote-core (>=5.0.0,<6.0.0)",
  # "pyannote-metrics (>=3.2.1,<4.0.0)",
  "pyaudio (>=0.2.14,<0.3.0)",
  "pydantic-settings (>=2.10.1,<3.0.0)",
  # "webdataset (>=1.0.2,<2.0.0)"
  "python-docx (>=1.2.0,<2.0.0)",
  "rapidfuzz (>=3.14.0,<4.0.0)",
  "ruff (>=0.12.11,<0.13.0)",
  "sqlalchemy (>=2.0.43,<3.0.0)",
  # "scipy (>=1.16.1,<2.0.0)",
  # "sentencepiece (>=0.2.1,<0.3.0)",
  # "soundfile (>=0.13.1,<0.14.0)",
  # "torch (>=2.8.0,<3.0.0)",
  # "torchaudio (>=2.8.0,<3.0.0)",
  # "transformers (>=4.56.0,<5.0.0)",
  "uvicorn (>=0.35.0,<0.36.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Set the maximum line length to 120 characters
line-length = 120

# Target Python 3.12
target-version = "py312"

# Exclude files/directories
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    ".env",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "Q",   # flake8-quotes
]

# Ignore specific rules if needed
ignore = [
    "E501",  # Line too long (handled by line-length setting)
    "B008",  # Do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
# Ignore specific rules for specific files
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py files

[tool.ruff.lint.isort]
# Import sorting configuration
known-first-party = ["paathguide"]
force-single-line = false
force-sort-within-sections = true
single-line-exclusions = ["typing"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Use spaces for indentation
indent-style = "space"
# Respect existing line endings
line-ending = "auto"
